Refactor external
