+++ network/src/external/channel.rs
@@ -25,8 +25,8 @@ use snarkos_errors::network::ConnectError;
 use std::{net::SocketAddr, sync::Arc};
 use tokio::{io::AsyncWriteExt, net::TcpStream, sync::Mutex};
 
-/// A Channel for reading and writing messages to a peer.
-/// Storing two streams allows for simultaneous reading/writing.
+/// A channel for reading and writing messages to a peer.
+/// The channel manages two streams to allow for simultaneous reading and writing.
 /// Each stream is protected by an Arc + Mutex to allow for channel cloning.
 #[derive(Clone, Debug)]
 pub struct Channel {
