+++ network/src/external/message_types/pong.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{
+use crate::external::{
     message::{Message, MessageName},
     message_types::Ping,
 };
