+++ network/src/external/protocol/ping/ping.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{
+use crate::external::{
     message_types::{Ping, Pong},
     Channel,
 };
